requirements.txt
Flask==2.2.1
keras==2.13.1
numpy
Pillow
tensorflow==2.13.1
pre-commit==4.0.1
pylint==3.3.1
pytest==8.3.3
Werkzeug==2.2.2